@@ -1,8 +1,25 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.17'
-gem 'sqlite3'
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '4.2.0'
+
+# Use postgresql as the database for Active Record
+gem 'pg', '~> 0.21.0'
+
+# Use SCSS for stylesheets
+gem 'sass-rails', '~> 5.0'
+# gem 'sass-rails', '~> 4.0.3'
+# Use Uglifier as compressor for JavaScript assets
+gem 'uglifier', '>= 1.3.0'
+
+# Use jquery as the JavaScript library
 gem 'jquery-rails'
+gem 'pry'
+
+gem 'protected_attributes' # https://github.com/rails/protected_attributes
+gem 'actionpack-action_caching' # https://github.com/rails/actionpack-action_caching
+gem 'activerecord-session_store' # https://github.com/rails/activerecord-session_store
+gem 'rails-observers' # https://github.com/rails/rails-observers
 
 group :test, :development do
   gem 'rspec-rails'
