@@ -13,4 +13,4 @@
 //= require jquery
 //= require jquery_ujs
 //= require bootstrap.min
-//= require_self
+//= require_tree .
