@@ -2,21 +2,37 @@ class ListsController < ApplicationController
   # GET /lists
   def index
     @lists = List.all
+    respond_to do |format|
+      format.html { render :index }
+      format.json { render json: @lists, status: 201 }
+    end
   end
 
   # GET /lists/1
   def show
     @list = List.find(params[:id])
+    respond_to do |format|
+      format.html { render :show }
+      format.json { render json: @list, status: 201 }
+    end
   end
 
   # GET /lists/new
   def new
     @list = List.new
+    respond_to do |format|
+      format.html { render :new }
+      format.json { render json: @list, status: 201 }
+    end
   end
 
   # GET /lists/1/edit
   def edit
     @list = List.find(params[:id])
+    respond_to do |format|
+      format.html { render :edit }
+      format.json { render json: @list, status: 201 }
+    end
   end
 
   # POST /lists
@@ -24,9 +40,15 @@ class ListsController < ApplicationController
     @list = List.new(params[:list])
 
     if @list.save
-      redirect_to @list, notice: 'List was successfully created.'
+      respond_to do |format|
+        format.html { redirect_to @list, notice: 'List was successfully created.' }
+        format.json { render json: @list, status: 201 }
+      end
     else
-      render action: "new"
+      respond_to do |format|
+        format.html { render :new }
+        format.json { render json: @list, status: 201 }
+      end
     end
   end
 
@@ -35,9 +57,15 @@ class ListsController < ApplicationController
     @list = List.find(params[:id])
 
     if @list.update_attributes(params[:list])
-      redirect_to @list, notice: 'List was successfully updated.'
+      respond_to do |format|
+        format.html { redirect_to @list, notice: 'List was successfully updated.' }
+        format.json { render json: @list, status: 201 }
+      end
     else
-      render action: "edit"
+      respond_to do |format|
+        format.html { render :edit }
+        format.json { render json: @list, status: 201 }
+      end
     end
   end
 
@@ -46,6 +74,9 @@ class ListsController < ApplicationController
     @list = List.find(params[:id])
     @list.destroy
 
-    redirect_to lists_url
+    respond_to do |format|
+      format.html { redirect_to lists_url, notice: 'List was successfully destroyed.' }
+      format.json { render json: @list, status: 201 }
+    end
   end
 end
