@@ -1,6 +1,6 @@
 class Task < ActiveRecord::Base
   belongs_to :list
-  attr_accessible :complete, :name
+  attr_accessible :complete, :name, :due_date
 
   validates :name, presence: true
 end
